refactoring commands and api helpers for use in other tools.
Closes #10

* moved to setup.cfg
* moved to using tox + nose for testing
* data-oriented utility functions are organized under kojismokydingo package
* cli functions and command classes are under kojismokydingo.cli package
* adds list-cgs command for presenting a list of Content Generators
* adds content generator permissions under userinfo command
