+++ kojismokydingo/__init__.py
@@ -22,28 +22,52 @@ license: GPL v3
 """
 
 
-from __future__ import print_function
-
-import re
-import sys
-
-from abc import ABCMeta, abstractmethod
-from argparse import ArgumentParser
 from collections import OrderedDict
-from functools import partial
-from koji import convertFault, Fault, GenericError
-from koji_cli.lib import activate_session
-from os.path import basename
-from six import add_metaclass
-from six.moves import \
-    zip as izip, \
-    zip_longest as izip_longest
+from koji import convertFault, read_config, Fault, ClientSession
+from koji_cli.lib import activate_session, ensure_connection
+from six.moves import zip
 
 from .common import chunkseq
 
 
+class ManagedClientSession(ClientSession):
+    """
+    A koji.ClientSession that can be used as via the 'with' keyword to
+    provide a managed session that will handle login and logout. Also
+    takes care of loading the relevant profile configuration.
+    """
+
+    def __init__(self, profile="koji"):
+        conf = read_config(profile)
+        server = conf["server"]
+        super(ManagedClientSession, self).__init__(server, opts=conf)
+
+    def __enter__(self):
+        activate_session(self, self.opts)
+        return self
+
+    def __exit__(self, exc_type, _exc_val, _exc_tb):
+        self.logout()
+        if self.rsession:
+            self.rsession.close()
+            self.rsession = None
+        return (exc_type is None)
+
+
+class AnonClientSession(ManagedClientSession):
+    """
+    A koji.ClientSession that can be used as via the 'with' keyword to
+    provide a managed session that will handle login and logout. Also
+    takes care of loading the relevant profile configuration.
+    """
+
+    def __enter__(self):
+        ensure_connection(self)
+        return self
+
+
 class BadDingo(Exception):
-    complaint = "Something bad"
+    complaint = "Something bad happened"
 
     def __str__(self):
         orig = super(BadDingo, self).__str__()
@@ -52,6 +76,14 @@ class BadDingo(Exception):
 
 class NoSuchBuild(BadDingo):
     complaint = "No such build"
+
+
+class NoSuchChannel(BadDingo):
+    complaint = "No such builder channel"
+
+
+class NoSuchContentGenerator(BadDingo):
+    complaint = "No such content generator"
 
 
 class NoSuchTag(BadDingo):
@@ -70,149 +102,8 @@ class NoSuchPermission(BadDingo):
     complaint = "No such permission"
 
 
-class PermissionException(BadDingo):
+class NotPermitted(BadDingo):
     complaint = "Insufficient permissions"
-
-
-printerr = partial(print, file=sys.stderr)
-
-
-def _rpm_str_split(s, _split=re.compile(r"(~?(?:\d+|[a-zA-Z]+))").split):
-    """
-    Split an E, V, or R string for comparison by its segments
-    """
-
-    return tuple(i for i in _split(s) if (i.isalnum() or i.startswith("~")))
-
-
-def _rpm_str_compare(left, right):
-    left = _rpm_str_split(left)
-    right = _rpm_str_split(right)
-
-    for lp, rp in izip_longest(left, right, fillvalue=""):
-
-        # Special comparison for tilde segments
-        if lp.startswith("~"):
-            # left is tilde
-
-            if rp.startswith("~"):
-                # right also is tilde, let's just chop off the tilde
-                # and fall through to non-tilde comparisons below
-
-                lp = lp[1:]
-                rp = rp[1:]
-
-            else:
-                # right is not tilde, therefore right is greater
-                return -1
-
-        elif rp.startswith("~"):
-            # left is not tilde, but right is, therefore left is greater
-            return 1
-
-        # Special comparison for digits vs. alphabetical
-        if lp.isdigit():
-            # left is numeric
-
-            if rp.isdigit():
-                # left and right are both numeric, convert and fall
-                # through
-                lp = int(lp)
-                rp = int(rp)
-
-            else:
-                # right is alphabetical or absent, left is greater
-                return 1
-
-        elif rp.isdigit():
-            # left is alphabetical but right is not, right is greater
-            return -1
-
-        # Final comparison for segment
-        if lp == rp:
-            # left and right are equivalent, check next segment
-            continue
-        else:
-            # left and right are not equivalent
-            return 1 if lp > rp else -1
-
-    else:
-        # ran out of segments to check, must be equivalent
-        return 0
-
-
-def rpm_evr_compare(left_evr, right_evr):
-    """
-    Compare two (Epoch, Version, Release) tuples.
-
-    Returns  1 if left_evr is greater-than right_evr
-             0 if left_evr is equal-to right_evr
-            -1 if left_evr is less-than right_evr
-    """
-
-    for lp, rp in izip_longest(left_evr, right_evr, fillvalue="0"):
-        if lp == rp:
-            # fast check to potentially skip all the matching
-            continue
-
-        compared = _rpm_str_compare(lp, rp)
-        if compared:
-            # non zero comparison for segment, done checking
-            return compared
-
-    else:
-        # ran out of segments to check, must be equivalent
-        return 0
-
-
-class NEVRCompare(object):
-    """
-    An adapter for Name, Epoch, Version, Release comparisons of a
-    build info dictionary. Used by the nevr_sort_builds function.
-    """
-
-    def __init__(self, binfo):
-        self.build = binfo
-        self.n = binfo["name"]
-
-        evr = (binfo["epoch"], binfo["version"], binfo["release"])
-        self.evr = tuple(("0" if x is None else str(x)) for x in evr)
-
-
-    def __cmp__(self, other):
-        # cmp is a python2-ism, and has no replacement in python3 via
-        # six, so we'll have to create our own simplistic behavior
-        # similarly
-
-        if self.n == other.n:
-            return rpm_evr_compare(self.evr, other.evr)
-
-        elif self.n < other.n:
-            return -1
-
-        else:
-            return 1
-
-
-    def __eq__(self, other):
-        return self.__cmp__(other) == 0
-
-
-    def __lt__(self, other):
-        return self.__cmp__(other) < 0
-
-
-    def __gt__(self, other):
-        return self.__cmp__(other) > 0
-
-
-def nevr_sort_builds(build_infos):
-    """
-    Given a sequence of build info dictionaries, sort them by Name,
-    Epoch, Version, and Release using RPM's variation of comparison
-    """
-
-    return sorted(build_infos, key=NEVRCompare)
 
 
 def _bulk_load(session, loadfn, keys, size):
@@ -235,7 +126,7 @@ def _bulk_load(session, loadfn, keys, size):
             # print(key, file=sys.stderr)
             loadfn(key)
 
-        for key, info in izip(key_chunk, session.multiCall()):
+        for key, info in zip(key_chunk, session.multiCall()):
             # print(key, info, file=sys.stderr)
 
             if info:
@@ -306,157 +197,6 @@ def bulk_load_rpm_sigs(session, rpm_ids, size=100, results=None):
         results[key] = info
 
     return results
-
-
-@add_metaclass(ABCMeta)
-class SmokyDingo(object):
-
-    group = "misc"
-    description = "A CLI Plugin"
-
-    # this is necessary for koji to recognize us as a cli command
-    exported_cli = True
-
-    # set of permission names that can grant use of this command. None
-    # or empty for anonymous access. Checked in pre_handle
-    permission = None
-
-
-    def __init__(self, name):
-        self.name = name
-
-        # these will be populated once the command instance is
-        # actually called
-        self.goptions = None
-        self.session = None
-
-        # this is used to register the command with koji in a manner
-        # that it expects to deal with
-        self.__name__ = "handle_" + name.replace("-", "_")
-
-        # allow a docstr to be specified on subclasses, but if absent
-        # let's set it based on the group and description.
-        if getattr(self, "__doc__", None) is None:
-            self.__doc__ = "[%s] %s" % (self.group, self.description)
-
-
-    def parser(self):
-        """
-        Override to provide an ArgumentParser instance with all the
-        relevant positional arguments and options added.
-        """
-
-        invoke = " ".join((basename(sys.argv[0]), self.name))
-        return ArgumentParser(prog=invoke, description=self.description)
-
-
-    def validate(self, parser, options):
-        """
-        Override to perform validation on options values. Return value is
-        ignored, use parser.error if needed.
-        """
-
-        pass
-
-
-    def pre_handle(self, options):
-        """
-        Verify necessary permissions are in place before attempting any
-        further calls.
-        """
-
-        if self.permission:
-            session = self.session
-            userinfo = session.getLoggedInUser()
-            userperms = session.getUserPerms(userinfo["id"]) or ()
-
-            if not (self.permission in userperms or "admin" in userperms):
-                msg = "Insufficient permissions for command %s" % self.name
-                raise PermissionException(msg)
-
-
-    @abstractmethod
-    def handle(self, options):
-        """
-        Perform the full set of actions for this command.
-        """
-
-        pass
-
-
-    def __call__(self, goptions, session, args):
-        self.goptions = goptions
-        self.session = session
-
-        parser = self.parser()
-        options = parser.parse_args(args)
-
-        self.validate(parser, options)
-
-        try:
-            activate_session(session, goptions)
-
-            self.pre_handle(options)
-            return self.handle(options) or 0
-
-        except KeyboardInterrupt:
-            print(file=sys.stderr)
-            return 130
-
-        except GenericError as kge:
-            printerr(kge)
-            return -1
-
-        except BadDingo as bad:
-            printerr(bad)
-            return -2
-
-        except Exception:
-            # koji CLI hides tracebacks from us. If something goes
-            # wrong, we want to see it
-            import traceback
-            traceback.print_exc()
-            raise
-
-
-class AnonSmokyDingo(SmokyDingo):
-
-    group = "info"
-    permission = None
-
-
-    def __init__(self, name):
-        super(AnonSmokyDingo, self).__init__(name)
-
-        # koji won't even bother fully authenticating our session for
-        # this command if we tweak the name like this. Since
-        # subclasses of this are meant to be anonymous commands
-        # anyway, we may as well omit the session init
-        self.__name__ = "anon_handle_" + self.name.replace("-", "_")
-
-
-class AdminSmokyDingo(SmokyDingo):
-
-    group = "admin"
-    permission = "admin"
-
-
-class TagSmokyDingo(SmokyDingo):
-
-    group = "admin"
-    permission = "tag"
-
-
-class TargetSmokyDingo(SmokyDingo):
-
-    group = "admin"
-    permission = "target"
-
-
-class HostSmokyDingo(SmokyDingo):
-
-    group = "admin"
-    permission = "host"
 
 
 #
