+++ setup.cfg
@@ -1,0 +1,85 @@
+
+
+[metadata]
+name = kojismokydingo
+version = 0.9.0
+description = A collection of Koji command-line plugins
+author = Christopher O'Brien
+author_email = obriencj@gmail.com
+url = https://github.com/obriencj/koji-smoky-dingo
+license = GNU General Public License v3 (GPLv3)
+
+classifiers =
+  Intended Audience :: Developers
+
+
+[options]
+packages =
+  kojismokydingo
+  kojismokydingo.cli
+
+requires =
+  koji
+  six
+
+zip_safe = true
+
+
+[options.entry_points]
+koji_smoky_dingo =
+  affected-targets = kojismokydingo.cli.tags:AffectedTargets
+  bulk-tag-builds = kojismokydingo.cli.builds:BulkTagBuilds
+  check-hosts = kojismokydingo.cli.hosts:CheckHosts
+  client-config = kojismokydingo.cli.clients:ClientConfig
+  list-cgs = kojismokydingo.cli.users:ListCGs
+  list-imported = kojismokydingo.cli.builds:ListImported
+  latest-archives = kojismokydingo.cli.archives:LatestArchives
+  list-build-archives = kojismokydingo.cli.archives:ListBuildArchives
+  perminfo = kojismokydingo.cli.users:PermissionInfo
+  renum-tag-inheritance = kojismokydingo.cli.tags:RenumTagInheritance
+  swap-tag-inhertance = kojismokydingo.cli.tags:SwapTagInheritance
+  userinfo = kojismokydingo.cli.users:UserInfo
+
+
+[tox:tox]
+envlist = py26,py27,py36,py37,flake8
+
+
+[testenv]
+commands =
+  python -B -m nose
+
+deps =
+  koji
+  nose
+  six
+
+sitepackages = false
+
+
+[testenv:flake8]
+commands =
+  python -B -m flake8
+
+deps =
+  flake8
+
+
+[nosetests]
+where = tests
+verbosity = 2
+all-modules = 1
+
+
+[flake8]
+
+# E303 complains about more than one blank lines between methods in a class
+# E741 ambiguous variable name
+# F401 ambiguous variable name
+
+ignore = E303,E741,F401
+exclude = setup.py,setup-meta.py,build,dist,tests,todo,.git,.tox
+
+
+#
+# The end.
